HTTP cache: Refactoring: Extract KeyHasher class
